+++ aws-sdk-core/spec/seahorse/client/base_spec.rb
@@ -272,6 +272,7 @@ module Seahorse
             Plugins::ParamConversion,
             Plugins::ParamValidation,
             Plugins::RaiseResponseErrors,
+            Plugins::ResponseTarget,
           ])
         end