+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Unreleased Changes
 ------------------
 
+* Issue - Aws::S3 - Resolved an issue where calling `Aws::S3::Client#get_object`
+  with a block would yield error data for non-200 HTTP responses.
+  Resolves [GitHub issue #207](https://github.com/aws/aws-sdk-core-ruby/issues/207).
+
 * Issue - CloudFormation - Resolved an issue with `Aws::CloudFormation` resource
   interfaces for accessing associated "has" associations. Fixes
   [GitHub issue #209](https://github.com/aws/aws-sdk-core-ruby/issues/209).